backend/requirements.txt
# Core FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# AI and image processing
google-generativeai>=0.8.0
google-generativeai>=0.8.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
Pillow>=10.0.0

# Database and authentication
supabase>=2.0.0
postgrest>=0.13.0

# Environment and configuration
python-dotenv>=1.0.0

# HTTP client and utilities
httpx>=0.25.0
requests>=2.31.0

# Data validation and serialization
email-validator>=2.1.0

# Security and authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Optional: For production deployment
gunicorn>=21.2.0
